Fix/heatmap rendering coordinates

+++ fan_app_interface/lib/features/map/presentation/stadium_map_page.dart
@@ -1267,14 +1267,14 @@ class StadiumMapPageState extends State<StadiumMapPage>
 
       // Extrair coordenadas do cellId (cell_X_Y ou cell_L_X_Y)
       final parts = cellId.split('_');
-      int? x, y;
+      double? x, y;
 
       if (parts.length == 3 && parts[0] == 'cell') {
-        x = int.tryParse(parts[1]);
-        y = int.tryParse(parts[2]);
+        x = double.tryParse(parts[1]);
+        y = double.tryParse(parts[2]);
       } else if (parts.length == 4 && parts[0] == 'cell') {
-        x = int.tryParse(parts[2]);
-        y = int.tryParse(parts[3]);
+        x = double.tryParse(parts[2]);
+        y = double.tryParse(parts[3]);
       } else {
         return;
       }
